src/third_party/blink/web_tests/wpt_internal/webxr/navigator_xr_blocked_by_getVRDisplays.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// This behavior is not in the webxr spec, so this is an internal-only test.
promise_test((t) => navigator.getVRDisplays().then(() => {
  assert_true(navigator.xr == null);
}), "Test that access to navigator.xr is blocked once navigator.getVRDisplays has been called.");

</script>
